requirements.txt
absl-py==2.0.0
alembic==1.12.1
astunparse==1.6.3
backports.functools-lru-cache==1.6.6
biopython==1.81
blinker==1.6.3
cachelib==0.9.0
cachetools==5.3.2
caniusepython3==7.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.8
filelock==3.12.2
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Caching==2.1.0
Flask-Migrate==4.0.7
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gast==0.5.4
gevent==22.10.2
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.2
grpcio==1.60.0
gunicorn==22.0.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
ml-dtypes==0.2.0
mock==1.0.1
mpmath==1.3.0
networkx==2.6.3
nltk==3.8.1
numpy #==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
optional-django==0.1.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pysam==0.22.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
RUST==0.1.1
sacremoses==0.0.53
safetensors==0.4.1
seaborn==0.13.0
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.10.1
tensorboard==2.0.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tokenizers==0.15.0
torch==1.13.1
tqdm==4.66.1
transformers==4.17.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
